Fix routing for lab run and edit
