+++ app-frontend/src/app/pages/lab/edit/edit.module.js
@@ -3,8 +3,8 @@ import angular from 'angular';
 import LabEditController from './edit.controller.js';
 import labEditComponent from './edit.component.js';
 
-require('./edit.scss');
 
+require('./edit.scss');
 const labEditModule = angular.module('pages.lab.edit', []);
 
 labEditModule.component('labEdit', labEditComponent);
