+++ app-frontend/src/app/pages/lab/run/run.module.js
@@ -3,8 +3,8 @@ import angular from 'angular';
 import LabRunController from './run.controller.js';
 import labRunComponent from './run.component.js';
 
-
 require('./run.scss');
+
 const labRunModule = angular.module('pages.lab.run', []);
 
 labRunModule.component('labRun', labRunComponent);
