+++ app-frontend/src/app/pages/lab/run/run.scss
@@ -1,5 +1,33 @@
-.run-project-name {
-  flex: 1;
-  font-weight: bold;
-  color: #333;
+.preview-control-bar {
+  position: absolute;
+  top: 0;
+  right: 0;
+  left: 0;
+  padding: 16px;
+  pointer-events: false;
+  & > * {
+    pointer-events: true;
+  }
+  text-align: center;
+  z-index: 1100;
+}
+
+.preview-info-bar {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  left: 0;
+  padding: 16px;
+  pointer-events: false;
+  display: flex;
+  justify-content: center;
+  & > * {
+    pointer-events: true;
+  }
+  text-align: center;
+  z-index: 1100;
+  .split-container {
+    flex: 1;
+    text-align: center;
+  }
 }
